--- a/src/Jeffijoe.MessageFormat.Tests/MetadataGenerator/GeneratedPluralRulesTests.cs
+++ b/src/Jeffijoe.MessageFormat.Tests/MetadataGenerator/GeneratedPluralRulesTests.cs
@@ -30,7 +30,7 @@ namespace Jeffijoe.MessageFormat.Tests.MetadataGenerator
                         new KeyedBlock("other", "дня")
                     },
                     new FormatterExtension[0]);
-            var request = new FormatterRequest(new Literal(1, 1, 1, 1, new StringBuilder()), "test", "plural", null);
+            var request = new FormatterRequest(new Literal(1, 1, 1, 1, ""), "test", "plural", null);
             var actual = subject.Pluralize("uk", arguments, new PluralContext(Convert.ToDecimal(Convert.ToDouble(args[request.Variable]))), 0);
             Assert.Equal(expected, actual);
         }
@@ -55,7 +55,7 @@ namespace Jeffijoe.MessageFormat.Tests.MetadataGenerator
                         new KeyedBlock("other", "дня")
                     },
                     new FormatterExtension[0]);
-            var request = new FormatterRequest(new Literal(1, 1, 1, 1, new StringBuilder()), "test", "plural", null);
+            var request = new FormatterRequest(new Literal(1, 1, 1, 1, ""), "test", "plural", null);
             var actual = subject.Pluralize("ru", arguments, new PluralContext(Convert.ToDecimal(args[request.Variable])), 0);
             Assert.Equal(expected, actual);
         }
@@ -78,7 +78,7 @@ namespace Jeffijoe.MessageFormat.Tests.MetadataGenerator
                         new KeyedBlock("other", "days")
                     },
                     new FormatterExtension[0]);
-            var request = new FormatterRequest(new Literal(1, 1, 1, 1, new StringBuilder()), "test", "plural", null);
+            var request = new FormatterRequest(new Literal(1, 1, 1, 1, ""), "test", "plural", null);
             var actual = subject.Pluralize("en", arguments, new PluralContext(Convert.ToDecimal(args[request.Variable])), 0);
             Assert.Equal(expected, actual);
         }