… src/Jeffijoe.MessageFormat/Jeffijoe.MessageFormat.csproj
… src/Jeffijoe.MessageFormat/MessageFormatter.cs
… src/Jeffijoe.MessageFormat/Parsing/IFormatterRequestCollection.cs
… src/Jeffijoe.MessageFormat/Parsing/Literal.cs
… src/Jeffijoe.MessageFormat/Parsing/LiteralParser.cs
… src/Jeffijoe.MessageFormat/Parsing/PatternParser.cs
… src/Jeffijoe.MessageFormat/StringBuilderPool.cs