Noticed some low-hanging fruit when working on parsing CLDR.

- Use StringBuilderPool via [Microsoft.Extensions.ObjectPool](https://www.nuget.org/packages/Microsoft.Extensions.ObjectPool/)
- Add some net5.0+ specific optimizations
- Do not go through with unescaping if there is nothing to unescape
- Store Literal InnerText as string

Method | Perf Improvement | Memory Improvement |
|------------------------- |---------:|----------:|
FormatWithCache | 19% | 54%
FormatWithCacheEscaped | 8% | 46%
FormatWithoutCache | 15% | 54%
FormatWithoutCacheEscaped | 8% | 52%



Before
|                   Method |Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |---------:|----------:|----------:|-------:|------:|------:|----------:|
|          FormatWithCache | 3.956 us | 0.0553 us | 0.0517 us | 1.1444 |     - |     - |   3.52 KB |
|    FormatWithCacheEscaped| 3.127 us | 0.0619 us | 0.0662 us | 0.9308 |     - |     - |   2.85 KB |
|       FormatWithoutCache | 7.657 us | 0.1434 us | 0.1535 us | 2.1210 |     - |     - |   6.52 KB |
| FormatWithoutCacheEscaped | 5.133 us | 0.0398 us | 0.0372 us | 1.3809 |     - |     - |   4.25 KB |

After
|                   Method |     Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------- |--------:|----------:|----------:|-------:|------:|------:|----------:|
|          FormatWithCache | 3.196 us | 0.0185 us | 0.0173 us | 0.5226 |     - |     - |    1.6 KB |
|    FormatWithCacheEscaped | 2.883 us | 0.0132 us | 0.0123 us | 0.4959 |     - |     - |   1.52 KB |
|       FormatWithoutCache | 6.473 us | 0.0375 us | 0.0351 us | 0.9537 |     - |     - |   2.95 KB |
| FormatWithoutCacheEscaped | 4.710 us | 0.0182 us | 0.0161 us | 0.6561 |     - |     - |   2.02 KB |

Benchmark code
```csharp
    class Program
    {
        static void Main(string[] args)
        {
            BenchmarkRunner.Run<MessageFormatBenchmark>();
        }
    }

    [MemoryDiagnoser]
    [HtmlExporter]
    [CsvExporter]
    public class MessageFormatBenchmark
    {
        private readonly Dictionary<string, object> _params;
        private readonly string _text;
        private readonly string _textEscaped;
        private readonly MessageFormatter _cacheFormatter;
        private readonly MessageFormatter _noCacheFormatter;

        public MessageFormatBenchmark()
        {
            _textEscaped =
                "These '{'count'}' and thoses '{count}' ain''t not escaped, which makes a total of {count, plural, one {a single pair} other {'#'# (=#) pairs}} of escaped braces.";
            _text =
                @"You have {count, plural, 
                            zero {no notifications}
                            one {just one notification}
                            =42 {a universal amount of notifications}
                            other {# notifications}
                      }. Have a nice day!"; ;
            _params = new Dictionary<string, object> { { "count", 2 } };
            _cacheFormatter = new MessageFormatter(true, "ru");
            _noCacheFormatter = new MessageFormatter(false, "ru");
        }

        [GlobalSetup]
        public void SetUp()
        {
            _cacheFormatter.FormatMessage(_text, _params);
            _cacheFormatter.FormatMessage(_textEscaped, _params);
        }

        [Benchmark]
        public string FormatWithCache()
        {
            return _cacheFormatter.FormatMessage(_text, _params);
        }

        [Benchmark]
        public string FormatWithCacheEscaped()
        {
            return _cacheFormatter.FormatMessage(_textEscaped, _params);
        }

        [Benchmark]
        public string FormatWithoutCache()
        {
            return _noCacheFormatter.FormatMessage(_text, _params);
        }

        [Benchmark]
        public string FormatWithoutCacheEscaped()
        {
            return _noCacheFormatter.FormatMessage(_textEscaped, _params);
        }
    }
```