@kostya9 2021-04-27
Fixed the warning, analyzers are happy now